requirements_full.txt
# Core Flask Dependencies
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.6
click==8.3.0
itsdangerous==2.2.0
MarkupSafe==3.0.3
blinker==1.9.0

# Database
pymongo==4.15.3
dnspython==2.8.0

# Authentication & Security
Flask-Login==0.6.3
bcrypt==5.0.0
PyJWT==2.10.1

# Forms & Validation
Flask-WTF==1.1.1
WTForms==3.2.1
email-validator==2.0.0

# Database ORM
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.43
greenlet==3.2.4

# File Handling
Pillow==10.0.0
python-docx==1.2.0
docx==0.2.4
lxml==6.0.2

# PDF Generation
reportlab==4.4.4

# Environment
python-dotenv==1.0.0

# Server
gunicorn==21.2.0

# Utilities
charset-normalizer==3.4.3
idna==3.10
typing_extensions==4.15.0

# ============================================
# NEW DEPENDENCIES FOR ENHANCEMENTS
# ============================================

# Real-Time Communication
flask-socketio==5.3.4
python-socketio==5.9.0
eventlet==0.33.3

# Data Science & Analytics
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
plotly==5.15.0
seaborn==0.12.2
scipy==1.11.1

# Machine Learning & AI
tensorflow==2.13.0
scikit-learn==1.3.0
opencv-python==4.8.0
scikit-image==0.21.0
imageio==2.31.1

# Image Processing
Pillow==10.0.0  # Already included above

# Additional Utilities
python-dateutil==2.8.2
pytz==2023.3
